requirements.txt
aiohttp==3.10.4
alembic~=1.14.1
bcrypt==4.3.0
click==8.1.7
cryptography==43.0.1
loguru==0.7.2
pyyaml~=6.0.2
SQLAlchemy==2.0.40
uvloop==0.21.0; sys_platform != 'win32'
winloop==0.1.8; sys_platform == 'win32'
